+++ voxygen/src/audio/mod.rs
@@ -9,12 +9,13 @@ pub mod soundcache;
 
 use channel::{AmbientChannel, AmbientChannelTag, MusicChannel, MusicChannelTag, SfxChannel};
 use fader::Fader;
+use music::MusicTransitionManifest;
 use sfx::{SfxEvent, SfxTriggerItem};
 use soundcache::{OggSound, WavSound};
 use std::time::Duration;
 use tracing::{debug, error};
 
-use common::assets::AssetExt;
+use common::assets::{AssetExt, AssetHandle};
 use rodio::{source::Source, OutputStream, OutputStreamHandle, StreamError};
 use vek::*;
 
@@ -45,6 +46,8 @@ pub struct AudioFrontend {
     sfx_volume: f32,
     music_volume: f32,
     listener: Listener,
+
+    mtm: AssetHandle<MusicTransitionManifest>,
 }
 
 impl AudioFrontend {
@@ -90,6 +93,7 @@ impl AudioFrontend {
             sfx_volume: 1.0,
             music_volume: 1.0,
             listener: Listener::default(),
+            mtm: AssetExt::load_expect("voxygen.audio.music_transition_manifest"),
         }
     }
 
@@ -108,6 +112,9 @@ impl AudioFrontend {
             sfx_volume: 1.0,
             music_volume: 1.0,
             listener: Listener::default(),
+            // This expect should be fine, since `<MusicTransitionManifest as Asset>::default_value`
+            // is specified
+            mtm: AssetExt::load_expect("voxygen.audio.music_transition_manifest"),
         }
     }
 
@@ -152,14 +159,19 @@ impl AudioFrontend {
                 let existing_channel = self.music_channels.last_mut()?;
 
                 if existing_channel.get_tag() != next_channel_tag {
+                    let mtm = self.mtm.read();
+                    let (fade_out, fade_in) = mtm
+                        .fade_timings
+                        .get(&(existing_channel.get_tag(), next_channel_tag))
+                        .unwrap_or(&(1.0, 1.0));
+                    let fade_out = Duration::from_secs_f32(*fade_out);
+                    let fade_in = Duration::from_secs_f32(*fade_in);
                     // Fade the existing channel out. It will be removed when the fade completes.
-                    existing_channel
-                        .set_fader(Fader::fade_out(Duration::from_secs(2), self.music_volume));
+                    existing_channel.set_fader(Fader::fade_out(fade_out, self.music_volume));
 
                     let mut next_music_channel = MusicChannel::new(&audio_stream);
 
-                    next_music_channel
-                        .set_fader(Fader::fade_in(Duration::from_secs(12), self.music_volume));
+                    next_music_channel.set_fader(Fader::fade_in(fade_in, self.music_volume));
 
                     self.music_channels.push(next_music_channel);
                 }